pyproject.toml
[project]
name = "glimmung"
version = "0.1.0"
description = "Agent dispatcher: leasable host queue for self-hosted runner work."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "azure-cosmos>=4.7",
  "azure-identity>=1.19",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "ulid-py>=1.1",
  "sse-starlette>=2.1",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/glimmung"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
